[omnibox] Add feature param to allow overriding starter pack url.
This CL adds a feature param to the starter pack expansion to
allow overriding the url for the new starter pack entry. This is to
enable testing in pre-prod channels.

Unlike crrev.com/c/5302777, this approach allows overriding the url
at delivery, eliminating the need to modify the DB or any in-memory
copies of the TemplateURL.

Bug: 41494524
Change-Id: I5dd87746998a09d5a99d43d4c4cd91a30157e043
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5304214
Reviewed-by: Moe Ahmadi <mahmadi@chromium.org>
Commit-Queue: Angela Yoeurng <yoangela@chromium.org>
Reviewed-by: Justin Donnelly <jdonnelly@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1262897}