Fix location icon view crash when icon name is nullptr
While trying to reproduce the bug we noticed segfaults happening when the name of an icon was nulltptr but still had a vector icon.

Digging further into the return types we realized that ImageModel().GetVectorIcon().vector_icon()->name returns a const char* const. This is problematic because we defined the `icon_name` variable as a const std::string. This causes an implicit conversion from char* to std::string when we go to store the variable. If the name was nullptr we will get a segfault since we cannot store a null value in std::string.

For this fix we just used the const char* const return type to be safe. We could have also used std::string*.

Bug: 324665414
Change-Id: I5a9dacf8db0747892e67ded877ec17b863f35216
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5311080
Commit-Queue: Darryl James <dljames@chromium.org>
Reviewed-by: Shibalik Mohapatra <shibalik@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1262910}