ui: gfx: Add Transform::Preserves2dAffine
Adds a function similar to Transform::Preserves2dAxisAlignment but
relaxed to preserving affine-ness. It returns true if, for 2d rects on
the x/y plane, a matrix can be represented as a 2d affine transform on
the x/y plane.

This is useful for delegated compositing since DrawQuads local space is
on the x/y plane. This will let us promote more kinds of quads that
have transforms which are "weird but whose weirdness doesn't affect the
quad".

Change-Id: Ic72b7aa6ebf4c4acff5daad72f4d6d043c29d79b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5304241
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Commit-Queue: Michael Tang <tangm@microsoft.com>
Reviewed-by: danakj <danakj@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1262918}