[Tab Group] Multi thumbnail use isTabInTabGroup
In multi thumbnail code use isTabInTabGroup to determine group
membership and whether to use a multi-thumbnail. Note that we also
need to support this for PseudoTab in the case that the tab model is
not yet initialized.

Note that if we cleanup InstantStart we can also delete PseudoTab
which I'd love to do.

Bug: b/41496693
Change-Id: I7c7768187e0f6a02b8160238b238cf604cd8a080
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5300785
Reviewed-by: Sky Malice <skym@chromium.org>
Commit-Queue: Calder Kitagawa <ckitagawa@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1262964}