App Management: Directly use WebApp uninstall dialog on Web App Settings
The only user of Uninstall() in App Service on non-ChromeOS is the Web
App Settings page. It's simpler to remove the extra layer of abstraction
and just call the Web App uninstall method directly from Web App
Settings.

This means that Uninstall() is now a CrOS-only method on
AppServiceProxy.

Bug: b/317152362
Change-Id: I24c8b6f5d9ff42e4e57e394faf58c3d723c4be5f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5298482
Commit-Queue: Tim Sergeant <tsergeant@chromium.org>
Reviewed-by: Maggie Cai <mxcai@chromium.org>
Reviewed-by: Daniel Murphy <dmurph@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1262980}