[gUM] Defer device selection until after permission grant
Old behavior was that gUM selected a device before showing
the permission prompt with no way for the user to influence
what was selected. When the new flag is enabled it will
change to waiting to selecting a device until after the user
grants permission. At that point it selects the best fit
format from the user's most preferred device and returns that
to the site.

Change-Id: I0ca245be4ac265085c531106ed9ca4fa352cd4fc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5226176
Commit-Queue: Bryant Chandler <bryantchandler@chromium.org>
Reviewed-by: Guido Urdaneta <guidou@chromium.org>
Reviewed-by: Chris Bookholt <bookholt@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1262996}