[Rounded Windows] Change the radius of WindowMiniView
Change the radius of WindowMiniView to match the radius of
rounded windows. Originally, the radius of window was decided to be
16dps but later changed to 12dps. Overview had an overhaul before
this decision and hence the difference.

This CL also inline the constants defines in the header ensuring
a single instance. (https://abseil.io/tips/168)

This CL also does other small refactors:
* Gets rid of redundant kMiniWindowHeaderCornerRadius constant.
* Improves readability of WindowMiniView::RefreshHeaderViewRoundedCorners().

Bug: b/323435089
Change-Id: I9b07f6e2d828c641e25b0542535253af95892e99
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5259812
Reviewed-by: Sammie Quon <sammiequon@chromium.org>
Commit-Queue: Zoraiz Naeem <zoraiznaeem@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1263001}