soda: Update live caption section to match Chrome.
Update the live caption section for Chrome OS to match the settings in
chrome, without live translate.

The settings are updated to use the appropriate analogous constructs
for Chrome OS instead of the ones for Chrome.

Browser Unit tests are ported from the chrome side appropriately, and
integrated into browser tests for chromeos.

An end to end functionality test is shown, which includes a number of
config/settings/flags updates not included in this change.

https://photos.app.goo.gl/o34APiduBE8TJPmdA

Bug: b/257166700
Test: Unit tests pass, and end to end functional.
Change-Id: I8fc048d3b98700405b5eb815facf8be67822a745
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5289315
Reviewed-by: Evan Liu <evliu@google.com>
Reviewed-by: Wes Okuhara <wesokuhara@google.com>
Reviewed-by: Abigail Klein <abigailbklein@google.com>
Commit-Queue: Rob Schonberger <robsc@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1263055}