Ensure window goes to NORMAL state after resize-toggled
This CL ensures that a window goes to NORMAL state after it’s
resize-toggled.
Previously, we used `widget->Restore()` but it sets
`aura::client::kIsRestoringKey` during the operation, which makes ash
move the window to any state other than NORMAL depending on the window
state history.
So we want to explicitly trigger the state change to NORMAL. Please note
that this doesn’t “bypass” the client-controlled transition model.
Setting a kShowStateKey just notifies ash to emit WM_EVENT and that’s
handled asynchronously with ARC++ as usual.

BUG=b:325409397
BUG=b:322137433
BUG=b:325399059
TEST=ResizeUtilTest

Change-Id: Ica4b3629f6bd9960a22198f67e6794edd5d25dd1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5308026
Commit-Queue: Toshiki Kikuchi <toshikikikuchi@chromium.org>
Reviewed-by: Kazuki Takise <takise@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1263063}