media/gpu/chromeos: Migrate OnFrameConverted to FrameResource
This changes VideoDecoderPipeline::OnFrameConverted() to accept a
FrameResource instead of a VideoFrame. All of the callers are simply
using the VideoFrameResource wrapper, but that will change after
NativePixmap-backed FrameResources are available.

This is the first change in a series that migrate VideoDecoderPipeline
to use FrameResource.

Bug: b:277581596
Test: video.Play.*hw* tests pass on Volteer.
Test: media_unittests pass on Volteer.

Change-Id: Ief4c484468adbc0b3763f4c68b934106700352bc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5304575
Reviewed-by: Andres Calderon Jaramillo <andrescj@chromium.org>
Reviewed-by: Hirokazu Honda <hiroh@chromium.org>
Commit-Queue: Nathan Hebert <nhebert@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1263066}