History: Fix small visual regression in cr-checkbox styling.
In crrev.com/c/5304013 a call to remove the "unresolved" CSS attribute
from cr-checkbox was removed, but that call was actually still needed,
for the purposes of the <history-item> in chrome://history.

Fixing, by replacing with the native ':not(:defined)' CSS selector
instead, to properly style the element before it is actually upgraded to
a custom element.

Bug: 40943652
Change-Id: I63481190264e66d5bf6d6d978fe9ecee357ba072
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5311875
Reviewed-by: John Lee <johntlee@chromium.org>
Commit-Queue: Demetrios Papadopoulos <dpapad@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1263077}