[ios][TipsNotifications] Use PushNotification utils to get opt-in status
This change makes SetUpList use PushNotificationSettingsUtils to get the
opt-in status for the notifications, which is used to determine whether
the SetUpList Notifications item should be marked as complete.
Previously it was accessing the prefs directly, which is an
implementation detail of PushNotificationService that other code
should not depend on.

Bug: 325279788
Change-Id: Iae8716d85ae972f15ffb2188d762ee6a10d5c799
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5310414
Reviewed-by: Ali Juma <ajuma@chromium.org>
Commit-Queue: Scott Yoder <scottyoder@google.com>
Cr-Commit-Position: refs/heads/main@{#1263108}