Make clicks on screen capture notification open default handler
When the kFileNotificationRevamp flag is on, make clicks on the
notifications shown after performing screen capture (screenshot or
screen recording) open the default app instead of the files app. This
matches the behaviour of file download notifications.

To avoid confusion in the metrics, use a new enum for this action.

Bug: b/319023176
Change-Id: I93e440040b8fd65db99dc088bfe9c56849a5683d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5298380
Reviewed-by: Ahmed Fakhry <afakhry@chromium.org>
Reviewed-by: Michele Fan <michelefan@chromium.org>
Commit-Queue: Timothy Loh <timloh@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1263112}