[omnibox] Fixes reporting zero-suggest metrics in searchboxstats.
crrev.com/c/5280536 had a subtle bug where the internal ACResult's
session data was being swapped with the published ACResult in Sync and
LastAsync autocomplete passes. This resulted in that data to be lost
from the internal ACResult and not be set on the ACMatches in
ACController::UpdateSearchboxStats(). The unit tests also failed to
detect that. This CL fixes that bug, renames ACResult::Swap() to
ACResult::SwapMatchesWith(), reorganizes the code and adjusts the
comments for clarity, and improves the unit tests.

crrev.com/c/5286774 also had a subtle bug where zero_prefix_enabled
was being set to true in ACController::UpdateSearchboxStats() which
is reached when zero-prefix suggestions are shown, not zero-prefix
suggestions could've been shown as it is expected by this metric to
report. This CL also fixes that by setting zero_prefix_enabled earlier
in ACController::Start().

This CL also enables the feature by default to facilitate merging this
change (1 CL to merge instead of 2) to m123.

Bug: 1307142
Change-Id: I0d53f429ff8b39aa5b4257bbf4219e729d6822ed
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5309407
Reviewed-by: Justin Donnelly <jdonnelly@chromium.org>
Commit-Queue: Moe Ahmadi <mahmadi@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1263124}