Add PickerKeyEventHandler and basic plumbing for keyboard navigation.
For Picker keyboard navigation, we'd like to allow users to easily type
and modify their search query, as well as easily navigate and select
search results.

To do this, we plan to keep focus on the Picker search field and
override TextfieldController::HandleKeyEvent to forward keyboard
navigation events (e.g. arrow keys) to other Picker views where
necessary. (We also plan to "highlight" and draw fake focus indicators
onto these views so that they appear focused to the user, similar to
the behaviour of omnibox and app list search results.)

As a starting point, this CL adds PickerKeyEventHandler and basic
plumbing of the key events to relevant Picker views. For now, only the
enter key is handled and it just selects the first search result.
Further key event handling will be added in followups.

Bug: b:322900302, b:325192974
Change-Id: I82c360aa6bf6754a8d48f547c4a627b1fa2a153f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5307463
Reviewed-by: Darren Shen <shend@chromium.org>
Commit-Queue: Michelle Chen <michellegc@google.com>
Cr-Commit-Position: refs/heads/main@{#1263138}