CCA: Refactor PTZ control logic
Digital zoom is planned to be controlled using Mojo, which is different
from the existing way to control PTZ using MediaStream API. To handle
both ways of controlling PTZ using the same UI, we add PTZController
interface to let both approaches implement.

In this CL, we create an interface PTZController, and moves the existing
control to a new class MediaStreamPTZController.

Bug: b:315740659
Test: tast run $DUT camera.CCAUIPTZ
Change-Id: I9489e4303bc27c9abb1a47ab1bf017fe1072769f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5124893
Commit-Queue: Kam Kwankajornkiet <kamchonlathorn@chromium.org>
Reviewed-by: Chu-Hsuan Yang <chuhsuan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1263149}