[Office upload] Log metrics when cannot get Drive relative path
Previously when trying to open a Drive url and the relative path
couldn't be retrieved, no metrics were logged. This means there would
be leakage in the TaskResult and the OpenErrors for Drive.

Add a new enum OfficeDriveOpenErrors::kCannotGetRelativePath to handle
this end state.

Bug: b/325983221
Change-Id: Ic2888f1ec51b47f7477afefaabbc0e822533947d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5308875
Reviewed-by: Nigel Tao <nigeltao@chromium.org>
Reviewed-by: Jeremie Boulic <jboulic@chromium.org>
Commit-Queue: Cassy Chun-Crogan <cassycc@google.com>
Cr-Commit-Position: refs/heads/main@{#1263161}