Prevent crash calling setDeviceMetricsOverride on autoresize window
Emulation.setDeviceMetricsOverride goes through a series of calls
leading to WebFrameWidgetImpl::SetScreenInfoAndSize, which has a DCHECK
for frames with autoresize enabled. This change returns an early error
to avoid getting to the DCHECK and crashing.

Emulation.canEmulate already returns false in this case, so another
solution would be to simply block all commands in the domain. However,
most of the domain does work, so this felt unnecessarily restrictive.

This change is being made in anticipation of future changes that will
make extension popups (which have autoresize enabled) more consistently
show as pages in CDP, therefore increasing the likelihood commands are
called.

Bug: N/A
Change-Id: I180b1603ee07ff42442652a87f6ee07997f443ed
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5300251
Reviewed-by: Simon Zünd <szuend@chromium.org>
Commit-Queue: Oliver Dunk <oliverdunk@chromium.org>
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1263287}