[Quick Delete] Fix crash in tab closure
There is a crash in the TabModelImpl.closeMultipleTabs due to tabs
not in the model getting closed with QD. This could be due to the delay
between when the tabs are filtered (before clearing browsing data) to
when they are closed after clearing browsing data (Which could lead to
some delays). Any closure of the tabs or switching of the tab model
during this delay would result in a crash.

This change moves the tab lookup to right when the animation is
triggered/tabs will be closed.

Bug: b/324905035
Change-Id: I37b29a10ef7495d1f08b70258326e21a780403ee
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5307638
Reviewed-by: Christian Dullweber <dullweber@chromium.org>
Commit-Queue: Zaina Al-Mashni <zalmashni@google.com>
Cr-Commit-Position: refs/heads/main@{#1263304}