[UNO] Support Sign in paused - PeopleHandler
With Uno the primary account should be set/cleared with explicit user
action.

The people handler on signout:
- Shows Gaia logout tab
- If a primary account with sync still exists, it invokes sync.

This relies on:
- In case of signin primary account,
  `DiceResponseHandler::ProcessDiceSignoutHeader` will revoke  all
  tokens on signout.
- On revoke sync consent, the `SigninManager` will clear all accounts if
  the accounts in the Gaia cookie is empty (/Logout).

These assumptions will be broken on follow-up CLs with:
- `SigninManager` no longer created in the Full Uno experiment
- Signout from the web only invalidated the primary account.

This CL ensures on Signout from Chrome Setting, the primary account is
removed. On next reconciliation cycle, the `AccountReconcilor` will
match the refresh tokens to the accounts in the Gaia cookie. As the
\Logout + \ListAccounts are network requests, it is expected the account
reconcilor will run after the primary account is removed.

Bug: b/319402216
Change-Id: If8edfdf0d21d6e50ac102d6776144d74bd90a2f9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5307136
Commit-Queue: Monica Basta <msalama@chromium.org>
Reviewed-by: David Roger <droger@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1263325}