[PriceDrop] Show module on cold startup
Currently, the ShoppingPersistedTabDataService is not properly
initialized because the attempt of initialization happens too early
that the tab model may not be ready yet. This CL fixes this issue by
postponing the initialization until the tab model is ready. Based
on local testing, this fix doesn't lead to PriceChangeModule being
timeout and gets rejected by MagicStack.

Bug: 326149285
Change-Id: I8d1e2d1fc219718d232492c5861514f9efa262c8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5312080
Reviewed-by: Xi Han <hanxi@chromium.org>
Commit-Queue: Xi Han <hanxi@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1263326}