[Reland][iOS] Plus addresses settings integration
This is an initial, rough implementation in advance of the full
set of requirements. This is intended to support an initial small
experiment, nothing more. It accordingly omits icons, etc. in favor
of a minimal implementation.

This is a reland of crrev.com/c/5097387; the test in question has now
been updated, and a run of the ios-simulator-noncq bot passed.

Bug: 1467623, b/308603630
Change-Id: I8717a3b008292afc1ef9eeecd9bee56fde35dbbe
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5302022
Commit-Queue: Matt Reichhoff <mreichhoff@chromium.org>
Reviewed-by: Ginny Huang <ginnyhuang@chromium.org>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1263327}