[iOS] Add close item support in tab group view
Ensure the view is updated when an item is manually close by the user.

See video:
https://drive.google.com/file/d/1aZHZYc9PKn3pmW65a8abebVCjhw6si9r/view?usp=drive_link

This CL also reduce the usage of `GridViewControllerDelegate` and use
the mutator instead to avoid any dependency with the tab grid view
controller directly. The replaced function is not fully deleted as there
is still some dependency in inactive tabs.

Bug: 1501837, 1457146
Change-Id: Icb50786914b95165b85e3cca32ce411e1370201b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5272359
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Commit-Queue: Aliona Dangla <alionadangla@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1263333}