[DIPS] Move Redirect Heuristic into its own class.
This moves the redirect heuristic out of DIPSWebContentsObserver to a
new class, RedirectHeuristicTabHelper, which is also a
RedirectChainDetector::Observer.

To keep the diff small, I made the minimal changes to the browser tests
to allow them to be moved, but left them in
dips_bounce_detector_browsertest.cc. A future CL will move them to
//chrome/browser/tpcd/heuristics as well.

Bug: 40286530, 40883201
Change-Id: I8ad313ec4a168d9b9ebdbc5e3770994ac3f4eeb6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5311015
Commit-Queue: Ryan Tarpine <rtarpine@chromium.org>
Reviewed-by: Peter Kasting <pkasting@chromium.org>
Reviewed-by: Anton Maliev <amaliev@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1263354}