[mullhet] Store previewed fields in a vector.
This CL is a no-op from the preview functionality perspective.

This CL is a step towards adapting the PWM agent code to be able to
fill or preview any form field. The previewing function stops assuming
that the previewed field belongs to the password form. The previewed
field is appended to the persisted list of previewer fields. When the
form preview needs to be cleared, the information from list field list
is used.

Effectively this makes it possible to append any field to this previewed
fields list. The field will be cleared after the call to the
PasswordAutofillAgent::ClearPreviewedForm().

Bug: b:325464792
Change-Id: Ibad16380f09a7d994c7219836b30d800782ae69c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5307518
Commit-Queue: Timofey Chudakov <tchudakov@google.com>
Reviewed-by: Christoph Schwering <schwering@google.com>
Cr-Commit-Position: refs/heads/main@{#1263366}