[mullhet] Don't query cache in the clear preview.
PWM agent will eventually support previewing and filling any type of
fields. The previewed fields are stored in the corresponding private
member. This CL changes the behavior of the PWM agent to not check if
the passed field to clear preview belongs to the password form or not.

This has the following effect on the AF agent:
AutofillAgent::ClearPreviewedForm() will clear all fields previously
previewed by it. This is expected to have no functionality effect
because password and autofill forms can't be previewed simultaneously.
This change also saved 1 cache lookup in the PWM agent.

Change-Id: I2740eb75213d3d93d73c1db2f78d76a901c80795
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5307397
Commit-Queue: Timofey Chudakov <tchudakov@google.com>
Reviewed-by: Jihad Hanna <jihadghanna@google.com>
Cr-Commit-Position: refs/heads/main@{#1263376}