Fix strictmode UnbufferedIoViolation error for WebView shell
Tested the following steps:
1. Launch Shell Browser
2. Search for Test URL
3. Tap on any Link
4. Long press to select the text & Observe the behaviour

Strictmode violation is not reproducing before and after change
on Pixel 8.

The bug is produced on Redmi Note 12 Pro.
After this enters Canary, we will request the original bug reporter
to verify on their device, because we don't have that exact device.

Bug: 41494686
Change-Id: I08c3264503106ca18da50d02d3bb1e474591613f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5296638
Reviewed-by: Bo Liu <boliu@chromium.org>
Reviewed-by: Wayne Jackson Jr. <wbjacksonjr@chromium.org>
Commit-Queue: Ali Butt <alibu@google.com>
Cr-Commit-Position: refs/heads/main@{#1263381}