[//components/viz][Graphite] Use non-volatile promise images on Apple
Graphite non-volatile promise images can be used in Chromium if the
underlying Dawn texture will be the same across the sequence of scoped
accesses related to this promise image.

On Apple, this is feasible now that IOSurfaceImageBacking caches WGPU
textures on a per-usage basis. We expect this change to bring
performance benefits.

Bug: 41497036
Change-Id: Ic7330052d3e6c7bd34b41f241440f8de99f92943
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5281935
Reviewed-by: Sunny Sachanandani <sunnyps@chromium.org>
Commit-Queue: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1263384}